Samples of code worked on in the Unititled horror game
Event Manager was created to ensure a cut in code inter depedance as when working in a team it would have been easy to break each others code. By create a event manager we were able to fire off our own events and code without breaking pre-exisiting systems.

This event system can be seen utilised in audio system and event 8.

The detect interactable would find objects with the correct tags and see if the player could intereact with them and highlight the object if that was the case. This was done using the unity raycast system